fix: fixes error with xhr.responseXML on 'text' responseType
this is what I get in chrome.
"Uncaught DOMException: Failed to read the 'responseXML' property from 'XMLHttpRequest': The value is only accessible if the object's 'responseType' is '' or 'document' (was 'text')."
it pretty much explains itself. xhr.responseXML is not readable in some cases and that FF fix should be guarded with responseType condition.